backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.6",
    "fastapi>=0.135.3",
    "hkdf>=0.0.3",
    "jwcrypto>=1.5.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
